requirements.txt
# PDF 解析
PyMuPDF>=1.23.0

# LLM 交互
openai>=1.12.0
pydantic>=2.5.0

# 模板渲染
jinja2>=3.1.2
markdown>=3.5.0

# 浏览器自动化
playwright>=1.40.0

# 环境变量管理
python-dotenv>=1.0.0

# 图像处理
Pillow>=10.0.0

# 日志
colorlog>=6.8.0

# Web UI
gradio>=4.0.0
streamlit>=1.28.0

